+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/home/phone/WooPosPhoneCartScreen.kt
@@ -2,6 +2,8 @@ package com.woocommerce.android.ui.woopos.home.phone
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Modifier
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import com.woocommerce.android.ui.woopos.common.composeui.WooPosPreview
@@ -9,6 +11,7 @@ import com.woocommerce.android.ui.woopos.common.composeui.designsystem.WooPosThe
 import com.woocommerce.android.ui.woopos.common.composeui.isPreviewMode
 import com.woocommerce.android.ui.woopos.home.cart.WooPosCartScreen
 import com.woocommerce.android.ui.woopos.home.cart.WooPosCartScreenProductsPreview
+import com.woocommerce.android.ui.woopos.home.cart.WooPosCartState
 import com.woocommerce.android.ui.woopos.home.cart.WooPosCartViewModel
 
 @Composable
@@ -19,13 +22,35 @@ fun WooPosPhoneCartScreen(
 ) {
     if (isPreviewMode()) {
         WooPosCartScreenProductsPreview(modifier.fillMaxSize())
-    } else {
+        return
+    }
+    val state by viewModel.state.observeAsState()
+    state?.let { rawState ->
         WooPosCartScreen(
             modifier = modifier.fillMaxSize(),
-            viewModel = viewModel,
+            state = applyPhoneOverrides(rawState),
+            onUIEvent = viewModel::onUIEvent,
             onPhoneBack = onBack,
         )
     }
+}
+
+private fun applyPhoneOverrides(state: WooPosCartState): WooPosCartState {
+    val hasItems = state.body is WooPosCartState.Body.WithItems
+    val effectiveCheckoutButton = when {
+        !hasItems -> WooPosCartState.CheckoutButtonState.Invisible
+        state.checkoutButtonState == WooPosCartState.CheckoutButtonState.Disabled ->
+            WooPosCartState.CheckoutButtonState.Disabled
+        else -> WooPosCartState.CheckoutButtonState.Enabled
+    }
+    return state.copy(
+        toolbar = state.toolbar.copy(
+            backIconVisible = true,
+            isClearAllButtonVisible = hasItems,
+        ),
+        areItemsRemovable = hasItems,
+        checkoutButtonState = effectiveCheckoutButton,
+    )
 }
 
 @Composable
