--- WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/util/WooPosLayoutInfo.kt
@@ -1,12 +1,0 @@
-package com.woocommerce.android.ui.woopos.util
-
-import android.content.Context
-import com.woocommerce.android.ui.woopos.util.ext.isWooPosPhoneLayout
-import dagger.hilt.android.qualifiers.ApplicationContext
-import javax.inject.Inject
-
-class WooPosLayoutInfo @Inject constructor(
-    @ApplicationContext private val context: Context,
-) {
-    fun isPhoneLayout(): Boolean = context.isWooPosPhoneLayout()
-}
